fix body

+++ server/Client.hpp
@@ -11,15 +11,15 @@ enum ClientState
 {
 	READING,	  // --> the start state of the client
 	READING_BODY, // --> when i see "Content-Length"
-	PROCESSING,	  // --> when reading the request is done 
-	WRITING		  // --> after reading the body 
+	PROCESSING,	  // --> when reading the request is done
+	WRITING		  // --> after reading the body
 };
 
 class Client
 {
 	private:
 		int clientFd;
-		int len_body;
+		size_t len_body;
 		std::string reqBuff;
 		std::string resBuff;
 		ClientState state;
@@ -29,7 +29,7 @@ class Client
 			HttpRequest req;		// private or public ???????????????
 			RequestParser parser;	// private or public ???????????????
 			HttpResponse res;		// private or public ???????????????
-			
+
 			Client();
 			Client(int fd);
 			Client(const Client& other);
